ios: fix build — primitive Double bridges direct + libsqlite3 link + script always-out-of-date
Three fixes surfaced building the Phase 2 scaffold against the iPad
(A16) iOS 26.5 simulator. App now installs and renders end-to-end:
shared-module `MIN_CONFIDENCE_THRESHOLD` (0.60), `scoreConfidence`
returning 0.95 for a both-match and 0.50 for single-axis or no match,
with the threshold-gating color applied correctly in Swift.

1. ContentView.swift — drop `.doubleValue` accessors. The KMP→Swift
   bridge inlines top-level Kotlin function returns of `Double` and
   `const val Double` companion fields directly as Swift `Double`,
   not as `KotlinDouble` wrappers. Calling `.doubleValue` on a
   primitive Swift `Double` was the only compile error.

2. project.pbxproj — add `-lsqlite3` to OTHER_LDFLAGS on both Debug
   and Release. SQLDelight's Native driver depends on
   `sqliter-driver-cinterop-sqlite3`, which calls `sqlite3_step` /
   `sqlite3_reset` / etc. iOS ships `libsqlite3.tbd` in the SDK; we
   just need to ask for it. Same shape Android already has via the
   SQLDelight Android driver's transitive `androidx.sqlite:sqlite`
   dependency.

3. project.pbxproj — set `alwaysOutOfDate = 1` on the gradle Run
   Script build phase. Without it Xcode warned every build about the
   phase running unconditionally because it declared no input/output
   paths. Marking it explicitly always-out-of-date silences the
   warning while preserving the actual behavior — gradle decides
   whether the framework needs relinking based on its own incremental
   compilation tracking, not Xcode's file-stamp comparison.

Verified end-to-end: built the .app, installed it on the iPad (A16)
simulator, launched it, screenshotted the SwiftUI render. Phase 2
done.

+++ iosApp/Parachord/ContentView.swift
@@ -48,9 +48,9 @@ struct ContentView: View {
 
                 row(label: "MIN_CONFIDENCE_THRESHOLD", value: threshold)
                 Divider()
-                row(label: "Imagine / Lennon → Imagine / Lennon", value: correctMatch.doubleValue)
-                row(label: "Imagine / Lennon → Imagine / A Perfect Circle", value: wrongArtist.doubleValue)
-                row(label: "Imagine / Lennon → Teen Spirit / Nirvana", value: noMatch.doubleValue)
+                row(label: "Imagine / Lennon → Imagine / Lennon", value: correctMatch)
+                row(label: "Imagine / Lennon → Imagine / A Perfect Circle", value: wrongArtist)
+                row(label: "Imagine / Lennon → Teen Spirit / Nirvana", value: noMatch)
             }
             .padding()
             .background(Color(.systemGray6))
@@ -93,7 +93,7 @@ struct ContentView: View {
             Spacer()
             Text(String(format: "%.2f", value))
                 .font(.callout.monospacedDigit())
-                .foregroundStyle(value >= threshold.doubleValue ? .green : .red)
+                .foregroundStyle(value >= threshold ? .green : .red)
         }
     }
 }
